feature user-login and app.js - merged user-login and app.js to pass user prop information and lift user state
